+++ cvmfs/garbage_collection/garbage_collector_impl.h
@@ -14,13 +14,13 @@
 #include "util/string.h"
 
 template<class CatalogTraversalT, class HashFilterT>
-const unsigned int GarbageCollector<CatalogTraversalT,
-                                    HashFilterT>::Configuration::kFullHistory =
-  std::numeric_limits<unsigned int>::max();
+const uint64_t GarbageCollector<CatalogTraversalT,
+                                HashFilterT>::Configuration::kFullHistory =
+  std::numeric_limits<uint64_t>::max();
 
 template<class CatalogTraversalT, class HashFilterT>
-const unsigned int GarbageCollector<CatalogTraversalT,
-                                    HashFilterT>::Configuration::kNoHistory = 0;
+const uint64_t GarbageCollector<CatalogTraversalT,
+                                HashFilterT>::Configuration::kNoHistory = 0;
 
 template<class CatalogTraversalT, class HashFilterT>
 const time_t GarbageCollector<CatalogTraversalT,
@@ -91,9 +91,9 @@ void GarbageCollector<CatalogTraversalT, HashFilterT>::PreserveDataObjects(
     if (!oldest_trunk_catalog_found_)
       oldest_trunk_catalog_ = std::min(oldest_trunk_catalog_, mtime);
     if (configuration_.verbose) {
-      const int    rev   = data.catalog->revision();
+      const uint64_t rev = data.catalog->revision();
       LogCvmfs(kLogGc, kLogStdout | kLogDebug,
-               "Preserving Revision %d (%s / added @ %s)",
+               "Preserving Revision %" PRIu64 " (%s / added @ %s)",
                rev,
                StringifyTime(data.catalog->GetLastModified(), true).c_str(),
                StringifyTime(catalog_info_shim_.GetLastModified(data.catalog),
@@ -131,9 +131,10 @@ void GarbageCollector<CatalogTraversalT, HashFilterT>::SweepDataObjects(
 
   if (configuration_.verbose) {
     if (data.catalog->IsRoot()) {
-      const int    rev   = data.catalog->revision();
+      const uint64_t rev = data.catalog->revision();
       const time_t mtime = static_cast<time_t>(data.catalog->GetLastModified());
-      LogCvmfs(kLogGc, kLogStdout | kLogDebug, "Sweeping Revision %d (%s)",
+      LogCvmfs(kLogGc, kLogStdout | kLogDebug,
+               "Sweeping Revision %" PRIu64 " (%s)",
                rev, StringifyTime(mtime, true).c_str());
     }
     PrintCatalogTreeEntry(data.tree_level, data.catalog);
@@ -156,7 +157,8 @@ void GarbageCollector<CatalogTraversalT, HashFilterT>::SweepDataObjects(
     static_cast<float>(unreferenced_trees_);
   if (threshold > last_reported_status_ + 0.1) {
     LogCvmfs(kLogGc, kLogStdout | kLogDebug,
-             "      - %02.0f%%    %u / %u unreferenced revisions removed [%s]",
+             "      - %02.0f%%    %" PRIu64 " / %" PRIu64
+             " unreferenced revisions removed [%s]",
              100.0 * threshold, condemned_trees_, unreferenced_trees_,
              RfcTimestamp().c_str());
     last_reported_status_ = threshold;
